Removed ProxyAPI and ServerAPI since implementations are the same
