+++ Plan/common/src/main/java/com/djrapitops/plan/api/CommonAPI.java
@@ -19,14 +19,22 @@ package com.djrapitops.plan.api;
 import com.djrapitops.plan.api.data.PlayerContainer;
 import com.djrapitops.plan.api.data.ServerContainer;
 import com.djrapitops.plan.api.exceptions.database.DBOpException;
+import com.djrapitops.plan.data.plugin.HookHandler;
+import com.djrapitops.plan.data.plugin.PluginData;
 import com.djrapitops.plan.db.access.Query;
 import com.djrapitops.plan.db.access.queries.containers.ContainerFetchQueries;
 import com.djrapitops.plan.db.access.queries.objects.ServerQueries;
 import com.djrapitops.plan.db.access.queries.objects.UserIdentifierQueries;
+import com.djrapitops.plan.system.database.DBSystem;
+import com.djrapitops.plan.system.database.databases.operation.FetchOperations;
+import com.djrapitops.plan.system.database.databases.sql.operation.SQLFetchOps;
 import com.djrapitops.plan.utilities.uuid.UUIDUtility;
 import com.djrapitops.plugin.logging.L;
+import com.djrapitops.plugin.logging.console.PluginLogger;
 import com.djrapitops.plugin.logging.error.ErrorHandler;
 
+import javax.inject.Inject;
+import javax.inject.Singleton;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
@@ -37,15 +45,34 @@ import java.util.UUID;
  *
  * @author Rsl1122
  */
-public abstract class CommonAPI implements PlanAPI {
+@Singleton
+public class CommonAPI implements PlanAPI {
 
+    private final DBSystem dbSystem;
     private final UUIDUtility uuidUtility;
+    private final HookHandler hookHandler;
+    private final PluginLogger logger;
     private final ErrorHandler errorHandler;
 
-    CommonAPI(UUIDUtility uuidUtility, ErrorHandler errorHandler) {
+    @Inject
+    public CommonAPI(
+            DBSystem dbSystem,
+            UUIDUtility uuidUtility,
+            HookHandler hookHandler,
+            PluginLogger logger,
+            ErrorHandler errorHandler
+    ) {
+        this.dbSystem = dbSystem;
         this.uuidUtility = uuidUtility;
+        this.hookHandler = hookHandler;
+        this.logger = logger;
         this.errorHandler = errorHandler;
         PlanAPIHolder.set(this);
+    }
+
+    @Override
+    public void addPluginDataSource(PluginData pluginData) {
+        hookHandler.addPluginDataSource(pluginData);
     }
 
     @Override
@@ -73,8 +100,6 @@ public abstract class CommonAPI implements PlanAPI {
         }
     }
 
-    protected abstract <T> T queryDB(Query<T> query);
-
     @Override
     public PlayerContainer fetchPlayerContainer(UUID uuid) {
         return new PlayerContainer(queryDB(ContainerFetchQueries.fetchPlayerContainer(uuid)));
@@ -93,5 +118,18 @@ public abstract class CommonAPI implements PlanAPI {
     @Override
     public String getPlayerName(UUID playerUUID) {
         return queryDB(UserIdentifierQueries.fetchPlayerNameOf(playerUUID)).orElse(null);
+    }
+
+    @Override
+    public FetchOperations fetchFromPlanDB() {
+        logger.warn("PlanAPI#fetchFromPlanDB has been deprecated and will be removed in the future. Stack trace to follow");
+        for (StackTraceElement element : Thread.currentThread().getStackTrace()) {
+            logger.warn(element.toString());
+        }
+        return new SQLFetchOps(dbSystem.getDatabase());
+    }
+
+    private <T> T queryDB(Query<T> query) {
+        return dbSystem.getDatabase().query(query);
     }
 }
